Refactor GenericProvider.py -> generic_provider.py
