+++ pytest.ini
@@ -203,7 +203,7 @@ flake8-ignore =
     medusa/media/ShowNetworkLogo.py D100 D102 D200 D400
     medusa/media/ShowPoster.py D100 D102 D200 D400
     medusa/providers/exceptions.py D200 D400
-    medusa/providers/GenericProvider.py D100 D101 D102 D202 D205 D400 D401 E501 F841
+    medusa/providers/generic_provider.py D100 D101 D102 D202 D205 D400 D401 E501 F841
     medusa/providers/nzb/NZBProvider.py D100 D101 D102 D200 F401
     medusa/providers/torrent/TorrentProvider.py D100 D101 D102 D205 D400
     medusa/recompiled/__init__.py D104
