+++ tests/legacy/providers/nzb_provider_tests.py
@@ -21,7 +21,7 @@ from __future__ import print_function
 import os
 
 import medusa as app
-from medusa.providers.GenericProvider import GenericProvider
+from medusa.providers.generic_provider import GenericProvider
 from medusa.providers.nzb.NZBProvider import NZBProvider
 from six import iteritems
 from .generic_provider_tests import GenericProviderTests
